+++ Assignments/ass2/test/euclidean_vector/Constructors/size_initial_value_constructor.cpp
@@ -6,12 +6,12 @@
 
 TEST_CASE("size_constructor_magnitude_empty") {
     const auto vector = comp6771::euclidean_vector(0, 4);
-    REQUIRE(vector.dimensions() == 0);
+    REQUIRE(static_cast<size_t>(vector.dimensions()) == 0);
 }
 
 TEST_CASE("size_constructor_magnitude_implicit") {
     const auto vector = comp6771::euclidean_vector(10, 4);
-    REQUIRE(vector.dimensions() == 10);
+    REQUIRE(static_cast<size_t>(vector.dimensions()) == 10);
     // check all values defaulting to initial value of 4
     REQUIRE(std::all_of(vector.begin(), vector.end(),
                         [&](auto value) { return value == 4; }));
@@ -21,7 +21,7 @@ TEST_CASE("size_constructor_magnitude_explicit") {
     auto dimension = int{3};
     auto magnitude = double{3.24};
     auto vector_object_initialised = comp6771::euclidean_vector(dimension, magnitude);
-    REQUIRE(vector_object_initialised.dimensions() == 3);
+    REQUIRE(static_cast<size_t>(vector_object_initialised.dimensions()) == 3);
     REQUIRE(std::all_of(vector_object_initialised.begin(), vector_object_initialised.end(),
                         [&](auto value) { return value == magnitude; }));
 }
