+++ Assignments/ass2/test/euclidean_vector/Friends/equals.cpp
@@ -19,8 +19,8 @@
 TEST_CASE("equals_works_correctly_equal_vectors_empty") {
     const auto vector_one = comp6771::euclidean_vector(0);
     const auto vector_two = comp6771::euclidean_vector(0);
-    REQUIRE(vector_one.dimensions() == 0);
-    REQUIRE(vector_two.dimensions() == 0);
+    REQUIRE(static_cast<size_t>(vector_one.dimensions()) == 0);
+    REQUIRE(static_cast<size_t>(vector_two.dimensions()) == 0);
     auto equal = (vector_one == vector_two);
     REQUIRE(equal);
 }
@@ -32,8 +32,8 @@ TEST_CASE("equals_works_correctly_equal_vectors_empty2") {
     const auto vector_one = comp6771::euclidean_vector(0);
     const auto empty_vector = std::vector<double>();
     const auto vector_two = comp6771::euclidean_vector(empty_vector.begin(), empty_vector.end());
-    REQUIRE(vector_one.dimensions() == 0);
-    REQUIRE(vector_two.dimensions() == 0);
+    REQUIRE(static_cast<size_t>(vector_one.dimensions()) == 0);
+    REQUIRE(static_cast<size_t>(vector_two.dimensions()) == 0);
     auto equal = (vector_one == vector_two);
     REQUIRE(equal);
 }
@@ -42,9 +42,9 @@ TEST_CASE("equals_works_correctly_equal_vectors_empty2") {
 TEST_CASE("equals_works_correctly_equal_vectors_default_constructor") {
     const auto vector_one = comp6771::euclidean_vector();
     const auto vector_two = comp6771::euclidean_vector();
-    REQUIRE(vector_one.dimensions() == 1);
+    REQUIRE(static_cast<size_t>(vector_one.dimensions()) == 1);
     REQUIRE(vector_one[0] == 0);
-    REQUIRE(vector_two.dimensions() == 1);
+    REQUIRE(static_cast<size_t>(vector_two.dimensions()) == 1);
     REQUIRE(vector_two[0] == 0);
     auto equal = (vector_one == vector_two);
     REQUIRE(equal);
@@ -55,9 +55,9 @@ TEST_CASE("equals_works_correctly_equal_vectors") {
     const auto vector_one = comp6771::euclidean_vector(10, 5);
     const auto vector_two = comp6771::euclidean_vector(10, 5);
     // check vectors are equal first by checking properties of internals
-    REQUIRE(vector_one.dimensions() == 10);
+    REQUIRE(static_cast<size_t>(vector_one.dimensions()) == 10);
     REQUIRE(std::all_of(vector_one.begin(), vector_one.end(), [&](auto value) { return value == 5; }));
-    REQUIRE(vector_two.dimensions() == 10);
+    REQUIRE(static_cast<size_t>(vector_two.dimensions()) == 10);
     REQUIRE(std::all_of(vector_two.begin(), vector_two.end(), [&](auto value) { return value == 5; }));
     // check if the equals now does the same thing
     auto equal = (vector_one == vector_two);
@@ -68,9 +68,9 @@ TEST_CASE("equals_works_correctly_equal_vectors") {
 TEST_CASE("equals_works_correctly_equal_vectors_assigned1") {
     const auto vector_one = comp6771::euclidean_vector(10, 5);
     const auto vector_two = comp6771::euclidean_vector(vector_one);
-    REQUIRE(vector_one.dimensions() == 10);
+    REQUIRE(static_cast<size_t>(vector_one.dimensions()) == 10);
     REQUIRE(std::all_of(vector_one.begin(), vector_one.end(), [&](auto value) { return value == 5; }));
-    REQUIRE(vector_two.dimensions() == 10);
+    REQUIRE(static_cast<size_t>(vector_two.dimensions()) == 10);
     REQUIRE(std::all_of(vector_two.begin(), vector_two.end(), [&](auto value) { return value == 5; }));
     auto equal = (vector_one == vector_two);
     REQUIRE(equal);
@@ -80,9 +80,9 @@ TEST_CASE("equals_works_correctly_equal_vectors_assigned1") {
 TEST_CASE("equals_works_correctly_equal_vectors_assigned2") {
     const auto vector_one = comp6771::euclidean_vector(10, 5);
     const auto vector_two = vector_one;
-    REQUIRE(vector_one.dimensions() == 10);
+    REQUIRE(static_cast<size_t>(vector_one.dimensions()) == 10);
     REQUIRE(std::all_of(vector_one.begin(), vector_one.end(), [&](auto value) { return value == 5; }));
-    REQUIRE(vector_two.dimensions() == 10);
+    REQUIRE(static_cast<size_t>(vector_two.dimensions()) == 10);
     REQUIRE(std::all_of(vector_two.begin(), vector_two.end(), [&](auto value) { return value == 5; }));
     auto equal = (vector_one == vector_two);
     REQUIRE(equal);
@@ -92,9 +92,9 @@ TEST_CASE("equals_works_correctly_equal_vectors_assigned2") {
 TEST_CASE("equals_works_correctly_unequal_vectors") {
     const auto vector_one = comp6771::euclidean_vector(10, 5);
     const auto vector_two = comp6771::euclidean_vector(15, 5);
-    REQUIRE(vector_one.dimensions() == 10);
+    REQUIRE(static_cast<size_t>(vector_one.dimensions()) == 10);
     REQUIRE(std::all_of(vector_one.begin(), vector_one.end(), [&](auto value) { return value == 5; }));
-    REQUIRE(vector_two.dimensions() == 15);
+    REQUIRE(static_cast<size_t>(vector_two.dimensions()) == 15);
     REQUIRE(std::all_of(vector_two.begin(), vector_two.end(), [&](auto value) { return value == 5; }));
     auto equal = (vector_one == vector_two);
     REQUIRE(!equal);
