+++ Assignments/ass2/test/euclidean_vector/Friends/not_equals.cpp
@@ -11,8 +11,8 @@
 TEST_CASE("not_equals_works_correctly_equal_vectors_empty") {
     const auto vector_one = comp6771::euclidean_vector(0);
     const auto vector_two = comp6771::euclidean_vector();
-    REQUIRE(vector_one.dimensions() == 0);
-    REQUIRE(vector_two.dimensions() == 1);
+    REQUIRE(static_cast<size_t>(vector_one.dimensions()) == 0);
+    REQUIRE(static_cast<size_t>(vector_two.dimensions()) == 1);
     REQUIRE(vector_two[0] == 0);
     auto not_equal = (vector_one != vector_two);
     REQUIRE(not_equal);
@@ -22,13 +22,13 @@ TEST_CASE("not_equals_works_correctly_equal_vectors_empty") {
 TEST_CASE("not_equals_works_correctly_modifications") {
     auto vector_one = comp6771::euclidean_vector(10, 5);
     const auto vector_two = comp6771::euclidean_vector(vector_one);
-    REQUIRE(vector_one.dimensions() == 10);
+    REQUIRE(static_cast<size_t>(vector_one.dimensions()) == 10);
     REQUIRE(std::all_of(vector_one.begin(), vector_one.end(), [&](auto value) { return value == 5; }));
-    REQUIRE(vector_two.dimensions() == 10);
+    REQUIRE(static_cast<size_t>(vector_two.dimensions()) == 10);
     REQUIRE(std::all_of(vector_two.begin(), vector_two.end(), [&](auto value) { return value == 5; }));
     vector_one[0] *= 2;
     REQUIRE(vector_one[0] == 10);
-    REQUIRE(vector_one.dimensions() == 10);
+    REQUIRE(static_cast<size_t>(vector_one.dimensions()) == 10);
     bool updated_correctly = true;
     for (auto i = 0; i < vector_one.dimensions(); i++) {
         if (i == 0 and vector_one.at(i) == 10) {
@@ -48,9 +48,9 @@ TEST_CASE("not_equals_works_correctly_modifications") {
 TEST_CASE("not_equals_works_correctly_equal_vectors") {
     const auto vector_one = comp6771::euclidean_vector(10, 5);
     const auto vector_two = comp6771::euclidean_vector(10, 5);
-    REQUIRE(vector_one.dimensions() == 10);
+    REQUIRE(static_cast<size_t>(vector_one.dimensions()) == 10);
     REQUIRE(std::all_of(vector_one.begin(), vector_one.end(), [&](auto value) { return value == 5; }));
-    REQUIRE(vector_two.dimensions() == 10);
+    REQUIRE(static_cast<size_t>(vector_two.dimensions()) == 10);
     REQUIRE(std::all_of(vector_two.begin(), vector_two.end(), [&](auto value) { return value == 5; }));
     auto not_equal = (vector_one != vector_two);
     REQUIRE(!not_equal);
