+++ Assignments/ass2/test/euclidean_vector/Operations/copy_asignment.cpp
@@ -9,37 +9,37 @@
 
 TEST_CASE("copy_assignment_const") {
     const auto vector = comp6771::euclidean_vector();
-    REQUIRE(vector.dimensions() == 1);
+    REQUIRE(static_cast<size_t>(vector.dimensions()) == 1);
     REQUIRE(vector[0] == 0);
     const auto copy_vector = vector;
-    REQUIRE(copy_vector.dimensions() == 1);
+    REQUIRE(static_cast<size_t>(copy_vector.dimensions()) == 1);
     REQUIRE(vector[0] == 0);
 }
 
 TEST_CASE("copy_assignment_non_const") {
     const auto vector = comp6771::euclidean_vector();
-    REQUIRE(vector.dimensions() == 1);
+    REQUIRE(static_cast<size_t>(vector.dimensions()) == 1);
     REQUIRE(vector[0] == 0);
     auto copy_vector = vector;
-    REQUIRE(copy_vector.dimensions() == 1);
+    REQUIRE(static_cast<size_t>(copy_vector.dimensions()) == 1);
     REQUIRE(vector[0] == 0);
 }
 
 TEST_CASE("copy_assignment_changes") {
     auto vector = comp6771::euclidean_vector();
-    REQUIRE(vector.dimensions() == 1);
+    REQUIRE(static_cast<size_t>(vector.dimensions()) == 1);
     REQUIRE(vector[0] == 0);
 
     auto copy_vector = vector;
-    REQUIRE(copy_vector.dimensions() == 1);
+    REQUIRE(static_cast<size_t>(copy_vector.dimensions()) == 1);
     REQUIRE(copy_vector[0] == 0);
 
     copy_vector[0] = 1;
     // check changes went through
-    REQUIRE(copy_vector.dimensions() == 1);
+    REQUIRE(static_cast<size_t>(copy_vector.dimensions()) == 1);
     REQUIRE(copy_vector[0] == 1);
     // original stays unchanged!
-    REQUIRE(vector.dimensions() == 1);
+    REQUIRE(static_cast<size_t>(vector.dimensions()) == 1);
     REQUIRE(vector[0] == 0);
 
 }
